Separated genbot from a godfile into a project sturcture
