+++ .github/genbot/__main__.py
@@ -1,0 +1,5 @@
+"""Entry point for `python -m genbot`."""
+
+from genbot.cli import main
+
+main()
